+++ test/stack-test.sh
@@ -22,11 +22,10 @@ docker build \
   -t "$OUTPUT_IMAGE" \
   .
 
-LOAD_ENV_VARS="source .profile.d/heroku-tailscale-buildpack.sh"
-START_TAILSCALE="heroku-tailscale-start.sh && sleep 10"
+LOAD_ENV_VARS="source .profile.d/heroku-tailscale-buildpack.sh && sleep 10"
 PROXYCHAINS_WRAPPER="proxychains4 -f /app/vendor/proxychains-ng/conf/proxychains.conf"
 TEST_TAILNET_CONNECTION="curl --connect-timeout 5 hello.ts.net"
-TEST_COMMAND="$LOAD_ENV_VARS && $START_TAILSCALE && $PROXYCHAINS_WRAPPER $TEST_TAILNET_CONNECTION"
+TEST_COMMAND="$LOAD_ENV_VARS && $PROXYCHAINS_WRAPPER $TEST_TAILNET_CONNECTION"
 
 docker run \
     --rm \